app7/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-SocketIO
eventlet
werkzeug
llama-cpp-python  # or your LLaMA python binding
